tensorflow/requirements.txt
absl-py==2.3.1
astunparse==1.6.3
certifi==2026.2.25
charset-normalizer==3.4.6
filelock==3.19.1
flatbuffers==25.12.19
fsspec==2025.10.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.0
h5py==3.14.0
hf-xet==1.4.2
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
joblib==1.5.3
keras==3.10.0
libclang==18.1.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
networkx==3.2.1
numpy==2.0.2
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pillow==11.3.0
protobuf==6.33.6
Pygments==2.19.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.8.0
tqdm==4.67.3
transformers==4.57.6
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.7
wrapt==2.1.2
zipp==3.23.0
fastapi
uvicorn
